pyproject.toml
[tool.poetry]
name = "HoneyPot-AutoOpen"
version = "1.0.0"
description = "Automatically opens and notifies honeygain's honeypot"
authors = ["code-raisan <yuri@rainyan.net>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
pyHoneygain = "^0.3.5"
python-dotenv = "^0.21.1"
discord-webhook = "^1.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
